requirements.txt
requests~=2.27.1
wxpusher~=2.2.0
APScheduler~=3.10.4
pycryptodome~=3.20.0